+++ scripts/types/pf2e-types.js
@@ -14,9 +14,6 @@ export class PF2eConsumable {
 
     /** @type string */
     sourceId;
-    
-    /** @type boolean */
-    autoDestroy;
 
     /** @type boolean */
     isAmmunition;
@@ -26,4 +23,23 @@ export class PF2eConsumable {
 
     /** @type number */
     quantity;
+
+    /** @type PF2eConsumableSystem */
+    system;
+}
+
+export class PF2eConsumableSystem {
+    /** @type PF2eConsumableUses */
+    uses;
+}
+
+export class PF2eConsumableUses {
+    /** @type number */
+    value;
+    
+    /** @type number */
+    max;
+    
+    /** @type boolean */
+    autoDestroy;
 }